backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
psycopg[binary,pool]
